Factor out _jobs.py.
